Moved server services into their own folder
