+++ Lib/Terminal/Service/Server/DeleteServerService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Lib\Terminal\Service;
+namespace Lib\Terminal\Service\Server;
 
 use Lib\Terminal\Infrastructure\TerminalServerDbContext;
